adding reservation logic to proccess

+++ freedom-co/backend/routes/carts.py
@@ -26,35 +26,47 @@ def get_cart_from_user():
 
 @carts_bp.route('/api/carts/clear', methods=['DELETE'])
 @jwt_required()
-def clear_and_proccess():
+def clear_and_confirm():
     username = get_jwt_identity().get('username')
 
-    # Getting cart given an username and deleting
-    cart = Cart.query.filter_by(username=username).all()
-    for item in cart:
-        # Searching item and blocking (avoiding race condition)
-        store_item = StoreItem.query.filter_by(id=item.item_id).with_for_update().first()
-        clean_size = item.size.strip()
+    try:
+        # Retrieving the purchase record for the user
+        purchase = Purchase.query.filter_by(username=username, 
+                                            status=PurchaseStatus.PENDING).first()
+        if not purchase:
+            return jsonify({"message": "No pending reservation found for the user."}), 404
 
-        # todo check reservation
+        # Confirming purchase
+        purchase.status = PurchaseStatus.CONFIRMED
 
-        # Updating item quantities
-        if store_item and (clean_size in store_item.size_quantity_pairs):
-            store_item.size_quantity_pairs[clean_size] -= item.quantity
-        else:
-            return jsonify({"message": "Item not found."}), 404
+        # Retrieving items from the cart associated with the purchase
+        cart_items = Cart.query.filter_by(purchase_id=purchase.id).all()
+        if not cart_items:
+            return jsonify({"message": "No items found in the cart for this reservation."}), 404
 
-        # Deleting item from cart
-        db.session.delete(item)
+        # Clearing cart
+        for item in cart_items:
+            db.session.delete(item)
 
-    # Confirming
-    db.session.commit()
-    return jsonify({"message": "Cart cleared successfully."}), 200
+        # All fine
+        db.session.commit()
+        return jsonify({
+            "message": "Reservation confirmed and items removed from cart successfully.",
+            "purchase_id": purchase.id
+        }), 200
+
+    except Exception as e:
+        # Rollback and handle any exception
+        db.session.rollback()
+        return jsonify({"message": str(e)}), 500
 
 @carts_bp.route('/api/carts/checkout', methods=['GET'])
 @jwt_required()
 def checkout_from_user():
     username = get_jwt_identity().get('username')
+
+    if Purchase.query.filter_by(username=username, status=PurchaseStatus.PENDING).first():
+        return jsonify({"message": "A pending reservation already exists."}), 400
 
     try:
         # Getting cart items given the username
